requirements.txt
crewai
crewai_tools
python-dotenv
pypdf
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
requests
